Fixed warnings/errors in clib QA tests.
[git-p4: depot-paths = "//depot/openwatcom/": change = 32304]

+++ bld/clib/_dos/c/_dostest.c
@@ -199,7 +199,7 @@ void TestFileOperations( void )
     struct find_t       fileinfo;
     unsigned long       flag;
     unsigned            attribute;
-    unsigned short      date,time;
+    unsigned            date,time;
     char                buffer[] = "This is a test for _dos_write().";
     char                tmpbuff[80];
     unsigned            len;
